pyproject.toml
[project]
name = "langchain-doc-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "certifi>=2026.2.25",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "markdown>=3.10.2",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.43.0",
]
